refactor: address rubocop offenses
